requirements.txt
Django==2.2.5
cloudinary==1.24.0
dj-database-url==0.5.0
gunicorn==20.0.4
django-cors-headers==3.7.0
django-haystack==3.0
django-simple-search==1.0.2
django-taggit==0.23.0
django-utils-six==2.0
django-vote==2.2.0
elasticsearch==1.9.0
flake8==3.7.7
pycodestyle==2.5.0
pyelasticsearch==1.4.1
pyflakes==2.1.1
pyparsing==2.4.7
pytest-django==3.5.1
sqlparse==0.4.1
whitenoise==5.2.0
Whoosh==2.7.4
Unipath==1.0
Pillow==3.3.0
psycopg2==2.6.1
python-decouple==3
